requirements.txt
appnope==0.1.0
audioread==2.1.5
bleach==2.1.2
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
librosa==0.5.1
llvmlite==0.22.0
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numba==0.37.0
numpy==1.13.3
olefile==0.45.1
opencv-python==3.4.0.12
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
resampy==0.2.0
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sounddevice==0.3.9
terminado==0.8.1
testpath==0.3.1
torch==0.3.0.post4
torchfile==0.1.0
torchnet==0.0.1
torchvision==0.1.9
tornado==4.5.3
tqdm==4.19.6
traitlets==4.3.2
urllib3==1.22
visdom==0.1.7
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.4
